made fromTurtle() the constructor of Page
